requirements.txt
crewai[tools]>=0.86.0,<1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
hatchling>=1.21.0